Move UpdateChecker to utilities package
